productivity-agent/requirements.txt
discord.py
fastapi
uvicorn
sqlalchemy
apscheduler
python-dotenv
openai
